Refactor: variables globales

+++ packages/web/app/utils.server.ts
@@ -1,6 +1,7 @@
 import * as path from 'path';
 import * as fs from 'fs';
 import { SongIndexPatch, SongsExtras, SongsProcessor } from '@iresucito/core';
+import { LowSync, JSONFileSync } from 'lowdb';
 // @ts-ignore
 import send from 'gmail-send';
 
@@ -23,52 +24,59 @@ const NodeLister = async (path: string) => {
   });
 };
 
-console.log('utils.server.ts __dirname', __dirname);
+declare global {
+  var db: any;
+  var folderSongs: SongsProcessor;
+  var folderExtras: SongsExtras;
+  var mailSender: (...args: any[]) => Promise<void>;
+}
 
-let folderExtras: SongsExtras = new SongsExtras(
-  path.resolve(__dirname + '/../data'),
-  NodeExists,
-  NodeWriter,
-  NodeReader,
-  fs.promises.unlink
-);
-
-export const folderSongs = new SongsProcessor(
-  path.resolve(__dirname + '/../public/songs'),
-  NodeLister,
-  NodeReader
-);
-
-export const getdb = async () => {
-  // @ts-ignore
-  const lowdb = await import('lowdb');
-  const { LowSync, JSONFileSync } = lowdb;
+if (globalThis.db === undefined) {
   const dataPath = path.resolve(__dirname + '/../data');
   const dbPath = path.join(dataPath, 'db.json');
-  console.log('getdb', dbPath);
   var db = new LowSync(new JSONFileSync(dbPath));
   db.read();
   // @ts-ignore
   db.data ||= { users: [], tokens: [] };
   db.write();
-  return db;
-};
+  globalThis.db = db;
+}
+
+if (globalThis.folderSongs === undefined) {
+  globalThis.folderSongs = new SongsProcessor(
+    path.resolve(__dirname + '/../public/songs'),
+    NodeLister,
+    NodeReader
+  );
+}
+
+if (globalThis.folderExtras === undefined) {
+  globalThis.folderExtras = new SongsExtras(
+    path.resolve(__dirname + '/../data'),
+    NodeExists,
+    NodeWriter,
+    NodeReader,
+    fs.promises.unlink
+  );
+}
+
+if (globalThis.mailSender === undefined) {
+  globalThis.mailSender = send({
+    user: 'javier.alejandro.castro@gmail.com',
+    pass: process.env.GMAIL_PASSWORD,
+    subject: 'iResucito Web',
+  });
+}
 
 export async function readLocalePatch(): Promise<any> {
-  const exists = await folderExtras.patchExists();
+  const exists = await globalThis.folderExtras.patchExists();
   if (exists) {
-    const patchJSON = await folderExtras.readPatch();
+    const patchJSON = await globalThis.folderExtras.readPatch();
     return JSON.parse(patchJSON);
   }
 }
 
 export async function saveLocalePatch(patchObj: SongIndexPatch | undefined) {
   const json = JSON.stringify(patchObj, null, ' ');
-  await folderExtras.savePatch(json);
+  await globalThis.folderExtras.savePatch(json);
 }
-
-export const mailSender = send({
-  user: 'javier.alejandro.castro@gmail.com',
-  pass: process.env.GMAIL_PASSWORD,
-  subject: 'iResucito Web',
-});
